Fix Error in Constructor for FFTAnalysis Class
This commit changes the internal handling of non-positive values
(NP <= 0) in the constructor for the FFTAnalysis class.  To the
end-user, that case is still a parsing error for a .FFT line.
However, within the constructor, I set np_=4 for that case. This
prevents an error in the test FFT/bad_fft_line2 during the
FFTInterface creation, later in that constructor, when compiling
with icc for the stokhos-basker build, and then running on four
 processsors in parallel.

SON Bug 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327