Remove explicit Sacado includes from N_DEV_Diode.h
For Reasons, it is necessary to include N_DEV_Diode.h in any Xyce
device that implements a "D" device.  It would be great if that were
not true, but it is at the moment.

That means that N_DEV_Diode.h has to be installed in ${prefix}/include
when Xyce is installed, so that Verilog-A diode devices (those that
have module attributes identifying them as diodes) can work as "D"
devices in netlists.

Unfortunately, N_DEV_Diode.h was including Sacado_No_Kokkos.h, which
means that anyone using buildxyceplugin must *also* have a full
Trilinos install in the same place that it lived when Xyce itself was
built.  This is not ideal, as it is a goal (so far unreachable) to
distribute a completely usable shared-library-enabled Xyce as a stand
alone package to some customers.  Having plugins depend explicitly on
Trilinos headers and libraries is a bad thing for that.

This commit moves the include of Sacado_No_Kokkos.h and the one
typedef referencing Sacado into the .C file instead of the .h file.

Unfortunately, while this allows buildxyceplugin to stop needing paths
to Trilinos headers when building any of the Verilog-A "toys", that's
only because there are no BJTs in the toys.  There is still a problem
in the N_DEV_BJT.h file needed by any Verilog-A "Q" devices --- and
that one is not as trivially fixed as this.  So there is still work to
be done.

Gitlab-ex #76
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/76