Remove sacado dependency from BJT header
Issue #76. "Work out way to get ADMS plugins built without needing
Trilinos headers"

The level-1 BJT header is one of the headers sometimes needed to build
ADMS models.  (It is needed for BJT models).   It contained a single
trilinos dependency, which was Sacado AD library.

It only uses sacado to provide analytic parameter derivatives for
sensitivity analysis.  Otherwise, all the derivatives in the BJT  are
hand-coded.   It was not necessary for the sacado dependency to be in
the header file.  It was only truly needed in the *C file, so I've moved
it there.

There was a very minor complication in doing this.  For several
functions that I implemented for sensitivities in this device, I
implemented them as templated functions, and created an explicit
specialization of those functions for a sacado Fad type.  But, the
templated aspect of these functions is not used at all.  It was only
there b/c I imagined refactoring the device in this direction, but
never got around to it.   At this stage, I *probably* won't be doing
that refactor anyway, so it is best to remove the templated stuff.
Now, these functions are not templated, and they are only defined
in the *C file, the only  place they are used.