Remove Sacado_No_Kokkos.h include from N_DEV_BJT.h
In commit 760d7ca, Eric removed some function template declarations
from N_DEV_BJT.h in furtherance of the goals of issue #76, an attempt
to make Verilog-A plugins buildable without requiring access to
Trilinos headers and libraries.

While that commit did in fact remove the need for the include of
Sacado_No_Kokkos.h from the header itself (which forces all Verilog-A
BJT plugins to require Trilinos headers), it did not actually remove
the include.

The include is now moved over to the .C file, where it is actually
needed.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/76