Fixes Basker interface to enable building against new interface
The Basker class name was changed from Trilinos 12.12 to now to
allow for both the templated Basker and ShyLU-Basker solver to be
compiled inthe same binary.  The autoconf system now accurately
detects which interface is available, if Basker has been built in
Amesos2.  Depending on that answer, one of the interfaces is enabled.
If neither of the known interfaces build in the autoconf test,
then the Amesos2-Basker interface is turned off.