Adds new strategy for shared node assignment
The random_shuffle that was used for parallel assignment of shared
nodes was removed to address test failures caused by non-deterministic
node assignment.  For bug_507, which change forced all shared nodes
to be assigned to one processor, in parallel, resulting in a performance
issue.  A simple balanced strategy is being added to assign shared
nodes based on which processor has the least number of shared nodes.
This fixes bug_507, which is a weekly test.

Gitlab issue #149