Support MAXFREQ Qualifier for SNR for .MEASURE FFT
This commit adds support for the MAXFREQ qualifier for the SNR
measure type for FFT measure mode.  It also makes the calculation
of the SNR metric into a public function in the FFTAnalysis class,
so that function can be used by both .FFT and .MEASURE FFT lines.
Additional work is needed to support MAXFREQ and MINFREQ for the
SFDR measure type. The THD calculations might also be centralized
in the FFTAnalysis class.

SON Bugs 1280 and 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327