Prepend independent variable in external output
In commit 83e31e93 (Feb 2018) I added a function to the General
External Output wrapper class to "normalize" the output list --- that
means, if it's a transient or frequency-domain run and the output list
does not include the independent variable (TIME or FREQUENCY), then we
prepend it to the list.

But after adding that function in that commit (and refining it in
3694598, 76d008c, 021c39b, and 17a616e) I never actually inserted code
to CALL this function.

Thus, the app note for the general interface *says* we prepend TIME or
FREQUENCY if the user has omitted it from the list, but we don't,
actually.

This got pointed out to us in Github issue 21 and is the
subject of our internal gitlab-ex issue #153.

Inserting a call to normalizeVarList_() into the ExternalOutputWrapper
class's constructor is all that was needed to do the trick.

Obviously, this feature has never been tested, and I'll have to fix
that.

https://github.com/Xyce/Xyce/issues/21
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/153