Extend testing of external output capability
The external outputter was intended to be designed such that if a user
requests a transient or AC/HB run and specifies an external outputter
that doesn't have TIME or FREQ (respectively) in its output field
list, then we automatically prepend the appropriate independent
variable to the list of requested outputs.  This test checks if that
output appears *anywhere* in the user's requested list, and only
prepends if it is not found at all.

This feature was never tested, and true to form because it wasn't
tested, it didn't work.  There was a missing call to the function that
actually did the work.

Now there is such a call, and this commit adds modification to the
testGenCoup test harness in order to support new tests in the
Xyce_Regression repo.  In particular, the "-iotest1" and "-iotest3"
options have been expanded to generate "noindex" output files that do
not explicitly include the independent variable in the output list.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/153
https://github.com/Xyce/Xyce/issues/21