Force Simulator class to reset error counts
The ErrorHandlingPKG houses a single static map of error
counts (mapping error type to count).  Thus, allocating multiple
Simulator objects (or any of the derived classes) and trying to run
them simultaneously (say in different threads) will aggregate all
their error counts together.

Worse, even allocating Simulator objects and running them
one at a time isn't safe --- errors from the run of one simulation
wind up being counted in the next.

The result is that if one allocates a Simulator and tries to run a
netlist that exits with a fatal error (say, during parsing), then
allocates a second object and passes it a good netlist, the second one
will fail at its very first safeBarrier call --- usually before any
error message has been printed *and* showing a count of errors that
includes not only the ones printed by the previous try but an extra
MSG_FATAL error (the one that actually resulted from the previous try
dying).

This commit fixes only the aggregation issue, which it does by
clearing the error counts in the Simulator constructor.  This is not a
full fix for the problem --- that will require a complete redesign of
how errors are handled in Xyce, so that they don't use static data.

After this commit, allocating and running a Simulator object then
allocating and running a second one will not pollute the second run
with errors from the first.

Trying to run both simultaneously (say in separate threads) will still
allow them to interfere with each other.  But we haven't ever really
said that that is supposed to work.

This addresses the immediate concern that led to Github issue 19 (and
which was turned into gitlab-ex issue #150).

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/150
https://github.com/Xyce/Xyce/issues/19