Support MAXFREQ Qualifier for SNR for .MEASURE FFT
This commit adds support for the MAXFREQ qualifier for the THD
measure type for FFT measure mode.  It also makes the NBHARM
qualifier, for THD measures, work correctly if the FREQ qualifier
was used on the associated .FFT line.  It did that by making the
calculation of the THD metric into a public function in the
FFTAnalysis class, so that function can be used by both .FFT and
.MEASURE FFT lines. A public funtion to convert the THD value to dB,
while respecting the "noise floor" used by the FFTAnalysis class,
was also added for use by both .FFT and .MEASURE lines.

SON Bugs 1280 and 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327