Support MINFREQ Qualifier for SFDR for .MEASURE FFT
This commit adds support for both the MINFREQ and MAXFREQ qualifier
for the SFDR measure type for FFT measure mode.  However, only the
MINFREQ qualifier will be tested at this point.  A new function,
for calculating the SFDR for a .MEASURE line, was added to the
FFTAnalysis class, because the corresponding function for calculating
the SFDR for a .FFT line also sets the index where the frequency for
the SFDR was found.  This new function just returns the requested
SFDR value, in dB, without modifying any private variables in the
FFTAnalysis class.

SON Bugs 1280 and 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327