Refactors initial condition processing and setting
The initial condition code in the IO package was written to apply
to only transient analysis.  Other analysis types, that use block
linear algebra structures, could not apply initial conditions in
the correct way.  While the IO package should be processing the
.IC statements and making sure they are sane, applying them is the
responsibility of the linear algebra package.

This refactor simplifies the processing of .IC data and storage of
that data to only what is necessary.  That data is then passed to
non-member helper functions from the linear algebra package, along
with the linear system, to make sure the initial conditions are
correctly applied.  Similarly, the nonlinear solver uses the linear
system to ensure that the correct lids are used in the augmented
linear system calls.

Bug: 1172 (SON)