Allow Paths Relative to Directory of .INC and .LIB File
This commit changes how Xyce handles the file paths given on .INC
and .LIB lines.  Define the following:

  P1 = path relative to the directory of the file with the .INC
       (or .LIB) line.

  P2 = path relative to the directory of the top-level netlist.

In Xyce 7.2, and earlier, path P2 was the only option.  The new
algorithm is that path P1 is preferred.  However, P2 will be then
be used as a "fallback path" if it exists.  Absolute paths should
still be used and have precedence.  However, that constraint may
not always be enforced on Windows systems.  That limitation will
be addressed in future commits.

SON Bug 1325
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1325