Stop indiscriminately replacing underscores in rawfiles
Ever since commit c3b8b217 (2005), Xyce's "override" raw file output
has blindly replaced the last underscore in an internal name with "#"
when writing out the raw file header.

This happened only in the "override" operation of raw files, that is,
when Xyce is invoked with the "-r rawfilename" option --- in this
case, Xyce dumps the entire solution vector to the raw file, and the
header contains the list of internal names rather than the names one
would use on the .print line.

Because IC-CAP expects the branch current of a voltage source to
appear in SPICE3F5 raw files with the name "sourcename#branch" and
Xyce's internal name for this variable is "sourcename_branch", commit
c3b8b217 took a fairly lazy approach to "fixing" it for IC-CAP use ---
on output, the last underscore in any internal name is replaced with
"#".

This is bad.  Yes, it is necessary that "_branch" be replaced with
"#branch" for IC-CAP, but none of the other replacements were needed.
And in fact, it's absolutely *wrong* to do this to any node name that
a user might have created with underscores in the name --- the last
underscore will be changed to "#" incorrectly, and now the raw file
will contain nodes named something other than what the user expects.

This bit an open source user who was using "-r" to get rawfiles and
was seeing his node names scrambled.  In fact, that user really wanted
to use ".print format=RAW" (which doesn't munge anything, and outputs
headers with names exactly as requested on the .print line).  But the
error was still there.

This commit fixes the two override raw file outputters so that they
only replace "_branch" with "#branch" if the internal name *ends* with
"_branch".  This is case insensitive, so if some Xyce developer
created an internal branch current with "_BRANCH" in it instead of
"_branch" it will still work.

This change breaks three tests in the test suite, because those tests
had gold standards that contained variables that weren't branch
variables and whose internal names legitimately included underscores
--- these were inappropriately being munged into octothorpes, but we
didn't care so long as they were consistently being munged.  Now that
they're not, those tests failed.  This will be addressed by replacing
the gold standards.

Addresses #8
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/8