Remove incorrect maximum step limit in the delay device
The maximum step limit is supposed to be applied at the first step
out of a discontinuous point. However, the maximum step is set up when
there is a change in the slope. This happens when the first step
out of a break point is already accepted. So this maximum step does not
limit the first step out of a break point. Instead, it can limit
the second step which is not useful.

The maximum step is kind of applied for the first step using the
isConverged technique. But it is not optimal. It can lead to a
unnecessary small first step because device non-convergence is a
nonlinear failure which can lead to large cut of the time step.
This commit removes the incorrect maximum step limit. This change
does not affect the current simple tests. But it reduces the number
of time steps in more complex customer circuits.