Make .param use of specials(TEMP) less restrictive
Issue #157.
"Make the use of specials (TEMP, TIME, VT, FREQ) in .param less
restrictive"

This commit makes it so that this is now allowed (a .param INDIRECTLY
depending on a special variable like TEMP):

.options device temp=25
.global_param barney={TEMP}
.param fred={barney}

R1 1 0 {fred}
V1 1 0 5V

However, this is still not allowed (a .param DIRECTLY depending on a
special):

.options device temp=25
.param fred={temp}

R1 1 0 {fred}
V1 1 0 5V

I added a new function to the expression library, called
"getShallowSpecials".  This returns the state of special variables in
the expression, but only the ones that can be found via the initial
parsing of the expression, before any external dependencies have been
resolved.

Since the new expression library treats .param and .global param as the
same behind the scenes, this works.

Ultimately per bug 1190 in bugzilla, I want to make .param and
.global_param completely the same.  All the necessary changes for this
are (I think) in the parser, the N_UTL_Param class and the device
package.    Those parts of the code don't yet fully understand that
.param isn't really constant anymore.

I'm not sure at the moment how much work that will be in total.  So, in the
interim, I can take small steps in that direction, like this one.