Prevent max timestep function of delay device being called
In a recent commit 6fe9566eb, @tmei disabled the delay device's
setting of a max time step by making it return a large number (1e99)
unconditionally.  This will certainly work (it is what the function
did if it wanted to "unset" a short max time step), but it is not the
best solution now that it is determined the device shouldn't set a max
time step at all.

A better approach is simply to prevent the getMaxTimeStepSize from
ever being called in the first place.  This can be done in several
ways, from the easily undoable to the more permanent.

The easily undoable way is to have the device's "maxTimeStepSupported"
return false.  That's what I've done here.

The more permanent way is simply to remove both the
maxTimeStepSupported declaration/definition in the .h file *and* the
declaration of getMaxTimeStepSize in the .h file and its definition in
the .C file.  If it is determoned that it is no longer *ever* going to
be necessary to use this function, that's what should be done later.

Part of #122
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/