Improve .FFT Metric and .MEASURE FFT Calculations
This commit streamlines the code for the THD, SNR and SFDR calculations
to make more efficient use of the upper limit (FMAX), for both .FFT
metric and .MEASURE FFT calculations.  It also fixes the use of the
FMAX value for SNR calculations.  For the SNR, the end of the "signal
frequencies" is now at the first harmonic frequency if FMAX is <=
that frequency.

SON Bugs 1280 and 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327