Make buildxyceplugin define "__XYCE__" for ADMS
Some of our built-in models derived from Verilog-A have been modified
with Xyce-specific blocks inside ifdefs that check for definition of
"__XYCE__".  This is then either defined with a "`define" in the file
itself or by adding "-D__XYCE__" to the admsXml invocation in the
directory's Makefile.

Note that not all modified devices use this same symbol (which is
unimportant to most users):  some use "_XYCE_HACK_" (PSP) or
"__XYCE_COMPACT_MODELING" (VBIC).

But an open source user would like to emulate this trick and write his
own model with Xyce-specific blocks ifdefed with "__XYCE__" and have
buildxyceplugin take care of telling his model that it's for Xyce and
not some other system.

This is an eminently reasonable thing to ask for, and so I'm doing it.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/158