Remove the "step delay" device
The "step delay" device was a single-purpose experimental device
intended for use in an LDRD.  It turned out to be the wrong approach
and is being removed.

The device was never added to testing or documentation, so this commit
completely removes it (modulo another commit in a different repo
needed to keep the "DEV_PARAMS" working).

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/152