Handle Windows Path Relative to Dir of INC/LIB Files
This commit improves the handling of Windows paths for the
files given on .LIB and .INC lines.  It now explicitly determines
whether various Windows-style paths, for those files, are either
absolute paths, or paths that relative to the current working
directory.  The goal was to improve testability, by breaking that
logic out into separate helper functions.

SON Bug 1325
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1325