Add console for transient adjoint calculation
When you compute transient adjoints, the calculation requires a backward
time integration to be performed after the original forward solve is
complete.

Prior to today, when this calculation is performed, there was nothing
printed to the console output to indicate what was going on during this
stage.  There would appear to be a mysterious pause before the final
timing stats were output.

With this commit, I have added a minimal bit of text to be sent to
console output, to indicate that Xyce is performing the reverse time
integration associated adjoint sensitivities.