Document Xyce#158
Issue Xyce#158 called for the buildxyceplugin script to define a
unique symbol that a model author could use to isolate Xyce-specific
code and have that code properly compiled when a shared library plugin
is generated for it using the script.

This has been added to the "Xyce/ADMS Improvements" section of the
release notes.  Though it is associated with an issue in the issue
tracker, it isn't really a "fixed defect" but rather a new feature,
and so I have chosen to document it only in this section and not the
fixed defects table.

The issue is also documented in the Xyce/ADMS users guide.
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/158