Fix the max time step issue for the first step out of a break point
The maximum time step is ignored for the first step out of a break
point. The time step is determined differently when the simulation
is at a break point. In this case, the time integrator is re-initialized.
These are several factors that can limit the first step out of a
break point. But the maximum time step is not used to limit the step.
This commit fixes both maximum and minimum time step in this case.