Make .param specials use completely unrestrictive
Issue #157.
"Make the use of specials (TEMP, TIME, VT, FREQ) in .param less
restrictive"

I initially addressed this issue by allowing .param to depend on a
special reserved variable such as TEMP, but only if it was an
indirect dependence.  Today, I experimented with making it
completely unrestricted.  In other words, any .param can
depend on a special variable.

With the new expression library, .param is treated the same as
.global_param.  So it differs from the old library, which treated
.params as immutable constants after the parsing phase.  With respect to
this specific issue, I knew that the expression library could handle
"specials dependence".  I just wasn't sure how much of the rest of the
Xyce source needed to be modified to support it.  As it turns out, not
very much.  I just had to remove the error trap from the parser, which
is what I did for this commit.

There are two "error message" style tests which will now break.  They
will simply need to be excluded (or actually, removed).