Document gitlab issue 8
Issue 8 in our gitlab-ex repo was about the inappropriate, blind
substitution of "#" for the last "_" in any variable name when
outputting "override" raw files.  That was originially done in 2005 to
support IC-CAP use, but all IC-CAP ever cared about was that the
branch current for a voltage source had a name in a raw file that was
"source#branch" just like SPICE3F5 always did.  It doesn't care about
other names.

This was brought to our attention by an open source user who was
annoyed that his node names that contained underscores were being
mangled in the rawfile.  So the fix needs mention in the release
notes (and nowhere else).

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/8