Adds logic to output Amesos2 capability and correct solvers
The configuration tools, autoconf and cmake, detect whether KLU2
and Basker are available in Amesos2.  This commit uses that information
to print out the available solvers when Xyce is queried for its
capabilities.  It also adds some logic to the Amesos2 interface to
trap on whether a solver is being requested that is unavailable.