Update param section of the user guide
This is related to the following issues:

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/93
"Support TEMPER and HERTZ in the new expression library"A

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/157
"Make the use of specials (TEMP, TIME, VT, FREQ) in .param less restrictive."

Also this bugzilla issue:
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1190
"Re-think whether .PARAM statements need to evaluate to a constant"

As well as the new usage of parameters in Xyce that have been enabled by
the new expression library.    In this chapter, parameters  (both .param
and .global_param) and expressions were discussed at a high level.  The
text included many warnings and explanations about how the value of a
.param must be knowable at parse time, and how most expressions are
evaluating during parsing and only during parsing.    A lot of this text
isn't true any more so I've removed it.

At this point, .param and .global_param are *almost* the same, in that
.params are allowed to change during a simulation, are allowed to depend
directly on .global_params and can depend directly or indirectly on
special variables such as TEMP or TIME.

AFAIK, the only thing that a .param cannot do now is be
modified directly via .STEP, .DC, .SAMPLING, etc.  And, the code
isn't far away from supporting this feature.

The only thing holding back .param support in .STEP, etc is that the
database of parameters currently stored in the device package only
contains global params and not normal params.  If that data structure
were to be updated to include normal params, then this feature would
work for them as well.  But, this is *probably* the least important
aspect of params being non-constant.

The most important thing was to provide flexbility and
compatibility with other simulator netlists, which is mostly done.