Improve Handling of BINSIZ Qualifier for .MEASURE FFT
This commit improves the handling of the BINSIZ qualifier in the
Base:Base() constructor.  The binSize_ variable is now set correctly
there, although that value is not used yet in the SNR, SNDR or ENOB
measure calculations.  This is consistent with how the WEIGHT and
GOAL qualifiers are handled now.

SON Bug 1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280