Fix intel compiler detection in configure
Our configure was set up to detect that it is building with intel
compilers and if so it adds an RPATH to help our standard binaries
find intel libraries when they're bundled.

This detection was broken when we added
  AX_CXX_COMPILE_STDCXX(11,,mandatory)
to the configure, because that changes $CXX to append "-std=c++11" or
something else like it, and that's no longer "equal" to "icpc"

This commit makes the detection more of a wildcard, so if
it *contains* "icpc" it's detected as Intel.

In fact, it wouldn't even hurt if we unconditionally added this RPATH,
but at the moment I'm just bandaiding it so that it does what it was
originally conceived to do instead of nothing.