Fix intel detection for parallel
More problems with the code that sets RPATH, there was a block that
was looking to see if the CXX compiler was "mpicxx" or "mpic++" and if
so checking $MPI_CXX for intel compiler.  Thing is, the same
AX_CXX_COMPILE_STDCXX(11,,mandatory)
that was changing $CXX from "icpc" to "icpc -std=gnu++11" is also
changing it from "mpicxx" to "mpicxx -std=gnu++11".

This test needs to be a wildcard, too.

This issue broke our Intel compiler builds on Linux and OS X because
it was failing to insert proper RPATHs into the binary.