Minor cleanup to param handling
This commit is related to these two issues, one from gitlab and the
other from bugzilla.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/167
"Make globally-scoped .params available to the device package, so
globally scoped normal params will be equivalent to .global_param"

Also this bugzilla issue:
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1190
"Re-think whether .PARAM statements need to evaluate to a constant"

This commit can be thought of as "a lot of cleanup that I should have
done last year, but never had time to get around to"

One thing that this commit does is remove a pointless if-statement.
It was created back when I did the new expression refactor.  The code (as I
just noticed today) read:

if (condition)
  do something
else
  do something

Where both "do something" were exactly the same.   So, I've deleted the
1st, 3rd and 4th lines of the above statement, essentially.

This must have started out as something much more complicated, and once
I cleaned it up the original idea didn't hold up anymore, but I never
completely cleaned it up.

This commit also cleans up the CircuitContext::resolveParameter function
This is a minor change which moves 3 very large debug output blocks into
2 helper functions.  This makes the function a LOT more readable.  It
also does a very similar cleanup to the resolveParameterThatIsAdotFunc
function.

The resolveParameterThatIsAdotFunc function is in the CircuitContext, and
is very similar to the resolveParameter function.  Prior to the new expression
library, these 2 functions were simply one single
function that could optionally handle .func.    But during the new
expression refactor, it was easier for me if I separated it into 2.
Some aspects of the .func version allow for significant simplification.
I did not do that before, but have done that now.

Finally this commit also removes a fair bit of superfluous,
redundant and/or unneccessary code.