Add CMake-based installation instructions
This commit adds instuctions for building Xyce using the CMake system.
Those are in the "INSTALL.md" file. I tried to make it as easy as
possible, and made the assumption that people understand how to use
their own package managers. Most systems will have all the needed TPLs,
with the exception of Trilinos. If they need to build a TPL, themselves,
then they will need to figure that out, themselves (ADMS would be in
this category; but it is optional).

To simplify the Trilinos building, I created two CMake "initial cache"
files. These have a standard set of build directives for serial and
parallel builds of Trilinos, and can be pulled in using the CMake '-C'
flag.

There is a section on the "superbuild" process. I specifically describe
it as a "beta" capability, and encourage users to use the standard build
process. The standard process is so easy now, we may want to reconsider
the usefulness of the superbuid, anyway.

The INSTALL.md file instructs users to look in the INSTALL file for
instructions on using the autotools system. I also added the reverse
direction in the INSTALL file.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/1