change base class processParams to be benign
This is indirectly related to:

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/167
"Make globally-scoped .params available to the device package, so
globally scoped normal params will be equivalent to .global_param"

Also this bugzilla issue:
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1190
"Re-think whether .PARAM statements need to evaluate to a constant"

The base device instance class function "processParams" throws
a fatal error if you try to call it.  I think this function has been
like this since I first implemented it back in 2002.   A lot of other
base class functions used to throw fatal errors if they were called and
this helped motivate some code development from what I recall.

Anyway, I was experimenting a bunch with parameter handling today and
stumbled across this.  There are the occasional device instance classes
that don't implement this funciton.  One example is the Vcvs device,
which only has a single parameter, gain, and thus doesn't really need
process params.   But, if you attempt to do a .step sweep on
vcvs::gain, the code exits with error.

Anyway, this commit makes that error (and errors like it,
theoretically) stop.