Change name of "Globals" to UserDefinedParams
This is indirectly related to:

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/167
"Make globally-scoped .params available to the device package, so
globally scoped normal params will be equivalent to .global_param"

Also this bugzilla issue:
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1190
"Re-think whether .PARAM statements need to evaluate to a constant"

There has been a structure in the device package named "Globals".  It
contains 3 objects that have also had the word "globals" in their names
as well.

Per the bugs above, I'm attempting to make .global_params and .params be
more similar.  As part of that work, I am updating the name of this
structure and its contents.  It was kind of a klunky name to begin with,
and with my current work it was an incorrect description.  This is
a cosmetic change, but designed to help make the code more
understandable.