Add box truncation method for multi-tone HB
The current truncation method is based on a modified version of
diamond truncation in multi-tone HB, which is not a common method
for frequency mapping. This commit adds a box truncation method
for multi-tone HB. This will be used in the artificial frequency
mapping algorithm for HB.

The box truncation method supports unlimited number of
tones in HB. The resulting frequency spectrum is not sorted. The
mapped frequencies are sorted. But the real frequencies are not
sorted for the AFM algorithm for HB simulations.

Xyce#160