Improve the box truncation method to work with existing APFT algorithm
The box truncation method can lead to a frequency spectrum that is not sorted
and can also has duplicate frequencies. This is fine with the artificial
frequency mapping algorithm since the mapped frequencies are sorted and
unique. However, this does not work with the current Almost Periodic
Fourier Transform multi-tone HB method in Xyce. This commit improved
the box truncation so it can work with both techniques for HB

Xyce#160