Fixes various memory and numerical bugs in Basker solvers
Turns out the Basker solver does not delete the L and U factors in
the solver object, if the same solver object is used for multiple
solves.  The construction and deletion of the Basker object is now
handled such that a new solver object is created for each solve to
address this issue.  There are additional memory losses within the
solver, but those need to be fixed in Trilinos.

The BlockMatrixEntry had some incorrect calculations of multiplication
and division when the matrix types differed (diagonal and dense).
This was illuminated by the parallel testing, not because of the
parallelism, just because the devices were reordered which resulted
in a different order of operation in the factorization.  Nonlinear
circuits will have a combination of dense and diagonal blocks, so
the common_emitter test case is perfect for identifying issues in
the solver.

Xyce#172