Fix the period calculation for HB
The period calculation is straightforward for single tone HB. It is
not the same for multi-tone HB. The current way assumes the frequency
spectrum is sorted. This does not work with artificial frequency mapping
method. The actual frequencies are not sorted in this case. This commit
fixes the period calculation so it can handle general frequency spectrum.

Xyce#160