Create getMutableValue functions
This is related to:

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/167
"Make globally-scoped .params available to the device package, so
globally scoped normal params will be equivalent to .global_param"

Also this bugzilla issue:
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1190
"Re-think whether .PARAM statements need to evaluate to a constant"

There are many places in the code where .param is allowed and implicitly
.global_param is not allowed.

Many of these places rely upon the Param::getImmutableValue function.
This function attempts to return a value for a specified type.  If the
underlying type is different than the specified type, it converts it.
It also then resets the type.

If the underlying type is Util::EXPR, then the getImmutableValue
function will only succeed if the underlying expression is considered
"constant".  Currently it is not considered to be constant if it depends
on any global parameters.

So, any many parts of the code, calls to getImmutableValue prevents
things from depending on global_params.  This appears to be an attempt
to do error trapping, and prevent certain types of things from being
changeable.  Like, for example, values applied in a .IC statement, etc.

But having the test hinge on .global_param dependence isn't great.  If
using a global_param, but not updating it (via .STEP or something),
that should not be a problem.  The correct test would be to see if the
expression depends on something that actually changes.
This logic, and use of getImmutableValue is a blocker to making .param
and .global_param be the same things on the top level of the netlist.

Unfortunately, the param class did not contain a good alternative
function to use.  There is a getValue function, but it is strongly typed
and will throw an exception if you request the wrong type.

In this commit I have created an alternative set of functions, called
"getMutableValue".  Functionally they are similar to getImmutableValue,
except that they permit all expressions, and they do not set the type.