Make L,W params work with .global_param
This change is related to:

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/167
"Make globally-scoped .params available to the device package, so
globally scoped normal params will be equivalent to .global_param"

Also this bugzilla issue:
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1190
"Re-think whether .PARAM statements need to evaluate to a constant"

As part of this work, I've discovered that .param is allowed in many
more situations than .global_param.  In some cases, this was
intentional, and in others it is accidental.

The commit fixes a case that was accidental. It has to do with setting
some instance parameters.  The processing of L, W in the constructor of
this device happens before the function updateDependentParams is called.
As a result, if L or W depend on a global_param, they are wrong.

This commit re-arranges the code slightly, so this doesn't happen
anymore.  This affected the BUG_573 test case, if I changed the
CERTPARAM parameter to a global.