Update CMake build instructions & README
I've tried to streamline the CMake build instructions to keep them as
clean as possible. To that end, all the system-specific instructions
have been moved to a section at the end.

I also tried to make the README a little more cohesive by referring back
to the Sandia Xyce home page.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/1